Changed component by hook to simplify things
